--- a/client_log.txt
+++ b/client_log.txt
@@ -1,0 +1,4 @@
+Request failed: HTTPConnectionPool(host='localhost', port=8000): Max retries exceeded with url: /calculate (Caused by NewConnectionError('<urllib3.connection.HTTPConnection object at 0x7f7c505883d0>: Failed to establish a new connection: [Errno 111] Connection refused'))
+Request sent. Server response status: 500
+Request sent. Server response status: 500
+Request sent. Server response status: 200