--- a/client_multicore_log.txt
+++ b/client_multicore_log.txt
@@ -1,0 +1,1 @@
+Client request status: 200