--- a/server_log_final.txt
+++ b/server_log_final.txt
@@ -1,0 +1,23 @@
+INFO:     Started server process [3881]
+INFO:     Waiting for application startup.
+INFO:     Application startup complete.
+INFO:     Uvicorn running on http://127.0.0.1:8000 (Press CTRL+C to quit)
+Initializing NUTS using jitter+adapt_diag...
+Multiprocess sampling (2 chains in 2 jobs)
+NUTS: [p_a, p_b]
+
+                              Step      Grad      Sampli…
+  Progre…   Draws   Diverg…   size      evals     Speed     Elapsed   Remaini…
+ ──────────────────────────────────────────────────────────────────────────────
+  ━━━━━━━   3000    0         1.26      3         1802.86   0:00:01   0:00:00
+                                                  draws/s
+  ━━━━━━━   3000    0         1.08      3         1713.00   0:00:01   0:00:00
+                                                  draws/s
+
+Sampling 2 chains for 1_000 tune and 2_000 draw iterations (2_000 + 4_000 draws total) took 2 seconds.
+We recommend running at least 4 chains for robust computation of convergence diagnostics
+INFO:     127.0.0.1:47780 - "POST /calculate HTTP/1.1" 200 OK
+INFO:     Shutting down
+INFO:     Waiting for application shutdown.
+INFO:     Application shutdown complete.
+INFO:     Finished server process [3881]