--- a/server_log_multicore.txt
+++ b/server_log_multicore.txt
@@ -1,0 +1,31 @@
+INFO:     Started server process [3374]
+INFO:     Waiting for application startup.
+INFO:     Application startup complete.
+INFO:     Uvicorn running on http://127.0.0.1:8000 (Press CTRL+C to quit)
+Initializing NUTS using jitter+adapt_diag...
+Multiprocess sampling (2 chains in 2 jobs)
+NUTS: [p_a, p_b]
+Debug values: p1=0.1, p2=0.12, p_pool=0.11
+Input data: a_success=100, a_total=1000, b_success=120, b_total=1000
+SE term before sqrt: 0.00019580000000000002
+Final SE: 0.013992855319769444
+Frequentist calculation time: 0.0005 seconds
+
+                              Step      Grad      Sampli…
+  Progre…   Draws   Diverg…   size      evals     Speed     Elapsed   Remaini…
+ ──────────────────────────────────────────────────────────────────────────────
+  ━━━━━━━   3000    0         1.09      3         1547.88   0:00:01   0:00:00
+                                                  draws/s
+  ━━━━━━━   3000    0         1.63      1         1492.71   0:00:02   0:00:00
+                                                  draws/s
+
+Sampling 2 chains for 1_000 tune and 2_000 draw iterations (2_000 + 4_000 draws total) took 2 seconds.
+We recommend running at least 4 chains for robust computation of convergence diagnostics
+PyMC sampling time: 4.8041 seconds
+Post-sampling processing time: 0.0582 seconds
+Total function execution time: 4.8629 seconds
+INFO:     127.0.0.1:44010 - "POST /calculate HTTP/1.1" 200 OK
+INFO:     Shutting down
+INFO:     Waiting for application shutdown.
+INFO:     Application shutdown complete.
+INFO:     Finished server process [3374]