… __pycache__/main.cpython-310.pyc
… client_final_log.txt