…mpling.

I noticed the A/B test calculation endpoint (`/calculate`) was a bit slow because the MCMC sampling in PyMC was only using a single core.

I've updated the code to allow PyMC to use all available CPU cores for parallel chain execution.

When I tested this on an environment with 2 CPU cores, I saw a significant improvement:
- PyMC sampling time went down from about 44.4 seconds to roughly 4.8 seconds.
- The total time to execute the request also dropped from around 44.4 seconds to approximately 4.8 seconds.

I've also removed some temporary profiling and debug statements.